requirements.txt
streamlit
Pillow
mediapipe
openai
imageio
numpy
reportlab
